Fix constructor and pin-setup of Rotary Encoder
This fixes a major logical error. When the library
constructs the Rotary object, its constructors takes the
pin numbers and calls pinMode() and digitalWrite() on them.
This will initialize the MCU-local pins and might conflict
with other constructors trying to initialize that pin to something else.

The function has been refactored to not do that, since all we want
is to abuse the Rotary object logic for the state-table and rotation
detection.